Refactor and add loger for code review
